@@ -9,52 +9,56 @@
     "test": "ng test"
   },
   "private": true,
+  "engines": {
+    "node": ">=18.0.0"
+  },
   "dependencies": {
-    "@angular/animations": "~18.1.4",
-    "@angular/cdk": "^18.1.4",
-    "@angular/common": "~18.1.4",
-    "@angular/compiler": "~18.1.4",
-    "@angular/core": "~18.1.4",
-    "@angular/forms": "~18.1.4",
-    "@angular/material": "^18.1.4",
-    "@angular/platform-browser": "~18.1.4",
-    "@angular/platform-browser-dynamic": "~18.1.4",
-    "@angular/router": "~18.1.4",
-    "@azure/event-hubs": "^5.11.2",
-    "@azure/identity": "^4.3.0",
-    "@azure/openai": "^1.0.0-beta.12",
-    "@lottiefiles/lottie-player": "^2.0.2",
-    "@microsoft/applicationinsights-angularplugin-js": "^15.0.0",
-    "@microsoft/applicationinsights-web": "^3.0.4",
-    "axios": "^1.5.1",
+    "@angular/animations": "~19.2.10",
+    "@angular/cdk": "^19.2.15",
+    "@angular/common": "~19.2.10",
+    "@angular/compiler": "~19.2.10",
+    "@angular/core": "~19.2.10",
+    "@angular/forms": "~19.2.10",
+    "@angular/material": "^19.2.15",
+    "@angular/platform-browser": "~19.2.10",
+    "@angular/platform-browser-dynamic": "~19.2.10",
+    "@angular/router": "~19.2.10",
+    "@azure/event-hubs": "^6.0.0",
+    "@azure/identity": "^4.9.1",
+    "@azure/openai": "^2.0.0",
+    "@lottiefiles/lottie-player": "^2.0.12",
+    "@microsoft/applicationinsights-angularplugin-js": "^15.3.6",
+    "@microsoft/applicationinsights-web": "^3.3.7",
+    "axios": "^1.9.0",
     "crypto-js": "^4.2.0",
-    "crypto.js": "^3.1.2",
-    "ng-table-virtual-scroll": "^1.4.1",
-    "openai": "^4.52.0",
-    "rhea-promise": "^3.0.0",
-    "rxjs": "^7.8.1",
+    "crypto.js": "^3.3.4",
+    "ng-table-virtual-scroll": "^1.6.1",
+    "openai": "^4.98.0",
+    "rhea-promise": "^3.0.3",
+    "rxjs": "^7.8.2",
     "secure-web-storage": "^1.0.2",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.14.10"
+    "tslib": "^2.8.1",
+    "util": "^0.12.5",
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^18.0.0",
-    "@angular-devkit/build-angular": "~18.1.4",
-    "@angular/cli": "^18.1.4",
-    "@angular/compiler-cli": "~18.1.4",
-    "@types/crypto-js": "^4.1.3",
-    "@types/jasmine": "~3.8.0",
-    "@types/node": "^12.20.55",
+    "@angular-builders/custom-webpack": "^19.0.1",
+    "@angular-devkit/build-angular": "~19.2.11",
+    "@angular/cli": "^19.2.11",
+    "@angular/compiler-cli": "~19.2.10",
+    "@types/crypto-js": "^4.2.2",
+    "@types/jasmine": "~5.1.8",
+    "@types/node": "^22.15.18",
     "buffer": "^6.0.3",
-    "jasmine-core": "~3.8.0",
-    "karma": "^6.4.2",
+    "jasmine-core": "^5.7.1",
+    "karma": "^6.4.4",
     "karma-chrome-launcher": "^3.2.0",
     "karma-coverage": "^2.2.1",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
     "os-browserify": "^0.3.0",
     "path-browserify": "^1.0.1",
     "process": "^0.11.10",
-    "typescript": "~5.4.5"
+    "typescript": "^5.8.3"
   }
 }
