@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 import { LoggingService } from './services/logging.service';
 
 @Component({
+  standalone: false,
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
@@ -12,6 +13,6 @@ export class AppComponent {
 
   constructor(private router: Router,
     private logingService: LoggingService) {
-      
+
   }
 }
