@@ -17,6 +17,8 @@ export class EventHubSourceService implements Model.IFirewallSource {
     private model:Model.ModelService
     )  {
   }
+
+  private queueLength: number = 0;
   private defaultSleepTime: number = 1500;
   private consumerClient: EventHubConsumerClient | undefined;
   private subscription: any;
@@ -28,6 +30,7 @@ export class EventHubSourceService implements Model.IFirewallSource {
   public onMessageArrived?: ((message: string) => void);
 
   public async start() {
+    this.queueLength = await this.getQueueLenght();
 
     try {
       this.outputMessage(`connecting to azure event hub`);
@@ -59,85 +62,77 @@ export class EventHubSourceService implements Model.IFirewallSource {
             for (const event of events) {
               var eventBody: Model.EventHubBody = event.body;
 
-              // *********************** fix for malformed json - TO BE REMOVED WHEN FIXED IN AZURE
+              // track event for malformed json
               if (eventBody.records == null) {
-                console.log(`${eventBody.toString()}`);
-
-                var asString: string = event.body.toString();
-               
-                if (asString.endsWith('",}}]}')) {
-                  asString = asString.slice(0,-6) + '"}}]}' ;
-                  
-                  eventBody = JSON.parse(asString);
-                }
+                this.logginService.logEvent(`UNABLE TO PARSE eventBODY (malformed json): ${eventBody.toString()}`); 
               }
-              // *********************** end fix for malformed json
+              else {
+                for (const record of eventBody.records) {
+                  const resourceId:string = record.resourceId;
 
-              for (const record of eventBody.records) {
-                const resourceId:string = record.resourceId;
+                  if (resourceId && resourceId.includes("/PROVIDERS/MICROSOFT.NETWORK/AZUREFIREWALLS/") == true) {
+                    var row: Model.FirewallDataRow | undefined = undefined;
 
-                if (resourceId && resourceId.includes("/PROVIDERS/MICROSOFT.NETWORK/AZUREFIREWALLS/") == true) {
-                  var row: Model.FirewallDataRow | undefined = undefined;
+                    switch (record.category) {
+                      case "AzureFirewallNetworkRule":
+                      case "AzureFirewallApplicationRule":
+                      case "AzureFirewallDnsProxy":    {
+                        row = this.parseAzureFirewallRuleLegacy(record);
+                        break;
+                      }
+                      // new format (structured logs)
+                      case "AZFWDnsQuery":
+                      case "AZFWApplicationRule":
+                      case "AZFWNetworkRule":
+                      case "AZFWNatRule":
+                      case "AZFWIdpsSignature":
+                      case "AZFWThreatIntel": 
+                        row = this.parseAzureFirewallRule(record);
+                        break;
+                      default: {
+                        row = {
+                          rowid: this.getRowID(),
+                          time: record.time.toString(),
+                          category: "SKIPPED Category - " + record.category,
+                          protocol: "-",
+                          sourceip: "-",
+                          srcport: "-",
+                          targetip: "-",
+                          targetport: "-",
+                          action: "-",
+                          dataRow: record
+                        } as Model.FirewallDataRow;
 
-                  switch (record.category) {
-                    case "AzureFirewallNetworkRule":
-                    case "AzureFirewallApplicationRule":
-                    case "AzureFirewallDnsProxy":    {
-                      row = this.parseAzureFirewallRuleLegacy(record);
-                      break;
-                    }
-                    // new format (structured logs)
-                    case "AZFWDnsQuery":
-                    case "AZFWApplicationRule":
-                    case "AZFWNetworkRule":
-                    case "AZFWNatRule":
-                    case "AZFWIdpsSignature":
-                    case "AZFWThreatIntel": 
-                      row = this.parseAzureFirewallRule(record);
-                      break;
-                    default: {
-                      row = {
-                        rowid: this.getRowID(),
-                        time: record.time.toString(),
-                        category: "SKIPPED Category - " + record.category,
-                        protocol: "-",
-                        sourceip: "-",
-                        srcport: "-",
-                        targetip: "-",
-                        targetport: "-",
-                        action: "-",
-                        dataRow: record
-                      } as Model.FirewallDataRow;
-
-                      this.skippedRows++;
-                      this.onRowSkipped?.(this.skippedRows);
-                      break;
+                        this.skippedRows++;
+                        this.onRowSkipped?.(this.skippedRows);
+                        break;
+                      }
                     }
                   }
-                }
-                else {
-                  row = {
-                    rowid: this.getRowID(),
-                    time: record.time.toString(),
-                    category: "SKIPPED Res Type - " + resourceId,
-                  } as Model.FirewallDataRow;
+                  else {
+                    row = {
+                      rowid: this.getRowID(),
+                      time: record.time.toString(),
+                      category: "SKIPPED Res Type - " + resourceId,
+                    } as Model.FirewallDataRow;
 
-                  this.skippedRows++;
-                  this.onRowSkipped?.(this.skippedRows); 
-                }
-                moreRows++;
-                this.DATA.unshift(row);
+                    this.skippedRows++;
+                    this.onRowSkipped?.(this.skippedRows); 
+                  }
+                  moreRows++;
+                  this.DATA.unshift(row);
 
-                while (this.DATA.length > environment.EventsQueueLength) {
-                  this.DATA.pop();
+                  while (this.DATA.length > this.queueLength) {
+                    this.DATA.pop();
+                  }
                 }
-          
+
+                this.onDataArrived?.(this.DATA); 
               }
+              const messageArrived = moreRows + " more events received as of " + new Date().toLocaleString();
+              this.outputMessage(messageArrived);
 
-              this.onDataArrived?.(this.DATA); 
             }
-            const messageArrived = moreRows + " more events received as of " + new Date().toLocaleString();
-            this.outputMessage(messageArrived);
 
             // save last event for later resume
             this.lastEvent = events.pop();
@@ -193,9 +188,26 @@ export class EventHubSourceService implements Model.IFirewallSource {
     this.outputMessage("Logs successfully deleted!");
   }
 
+  private async getQueueLenght(): Promise<number> {
+    try {
+      const response = await fetch('/api/settings/local_queuelength');
+      
+      if (!response.ok) {
+        console.error('queue lenght API endpoint returned error:', response.status);
+        return 10;
+      }      
+      
+      // Parse the response as text and convert to number
+      const text = await response.text();
+      return Number(text);
+    } catch (error) {
+      console.error('Error fetching queue lenght endpoint:', error);
+      return 10;
+    }
+  }
+
   private outputMessage (text:string): void {
     this.onMessageArrived?.(text);
-    this.logginService.logTrace(text);
   }
 
   private parseAzureFirewallRuleLegacy(record: Model.AzureFirewallRecord): Model.FirewallDataRow {
